+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/modules/ModuleDockerSection.java
@@ -85,6 +85,7 @@ public class ModuleDockerSection implements ISectionBuilder {
         utils.addItem(diBuilder.getDocItem("docker/container-running-with-name?", false));
         utils.addItem(diBuilder.getDocItem("docker/container-start-by-name", false));
         utils.addItem(diBuilder.getDocItem("docker/container-stop-by-name", false));
+        utils.addItem(diBuilder.getDocItem("docker/container-remove-by-name", false));
         utils.addItem(diBuilder.getDocItem("docker/container-status-by-name", false));
         utils.addItem(diBuilder.getDocItem("docker/container-exec-by-name", false));
         utils.addItem(diBuilder.getDocItem("docker/container-has-log-msg", false));